Update CMakeLists.txt and Sprite class***

+++ include/tdl/Matrix/Transformation.hpp
@@ -16,22 +16,22 @@
             void setPosition(const Vector2u &position);
             void setPosition(u_int32_t x, u_int32_t y);
 
-            void setRotation(float angle);
+            void setRotation(double angle);
             void setScale(const Vector2f &factors);
-            void setScale(float factorX, float factorY);
+            void setScale(double factorX, double factorY);
             void setOrigin(const Vector2f &origin);
-            void setOrigin(float x, float y);
+            void setOrigin(double x, double y);
 
             const Vector2f &getPosition() const;
-            float getRotation() const;
+            double getRotation() const;
             const Vector2f &getScale() const;
             const Vector2f &getOrigin() const;
 
             void move(const Vector2f &offset);
-            void move(float offsetX, float offsetY);
-            void rotate(float angle);
+            void move(double offsetX, double offsetY);
+            void rotate(double angle);
             void scale(const Vector2f &factor);
-            void scale(float factorX, float factorY);
+            void scale(double factorX, double factorY);
 
             const Transform &getTransform() const;
 
@@ -39,7 +39,7 @@
             Vector2f m_origin;
             Vector2f m_position;
             Vector2f m_scale;
-            float m_rotation;
+            double m_rotation;
             mutable Transform m_transform;
             mutable bool TransformNeedUpdate;
         };
